Update: Convert to protocol buffer

+++ app/src/main/java/com/srilakshmikanthanp/clipbirdroid/types/enums/ErrorCode.kt
@@ -1,6 +1,6 @@
 package com.srilakshmikanthanp.clipbirdroid.types.enums
 
-enum class ErrorCode(val value: Byte = 0x00) {
+enum class ErrorCode(val value: Int = 0x00) {
   /**
    * Allowed Error Codes
    */
@@ -10,19 +10,13 @@ enum class ErrorCode(val value: Byte = 0x00) {
    * Companion Object
    */
   companion object {
-    /**
-     * Convert Byte to ErrorCode
-     */
-    fun fromByte(value: Byte): ErrorCode = when (value) {
+    fun fromByte(value: Int): ErrorCode = when (value) {
       InvalidPacket.value -> InvalidPacket
       InvalidPacket.value -> InvalidPacket
       else -> throw IllegalArgumentException("Invalid ErrorCode value: $value")
     }
 
-    /**
-     * Convert ErrorCode to Byte
-     */
-    fun toByte(status: ErrorCode): Byte {
+    fun toByte(status: ErrorCode): Int {
       return status.value
     }
   }
